cli cleanup and docs: move cli tools to separate cli package, with entrypoint still in warcio.cli move extractor and recompressor to separate cli package modules add readme description for extract tool
