+++ warcio/cli/checker.py
