+++ warcio/cli/indexer.py
